requirements.txt
tabulate
jsonschema
pyyaml